Support callback when visualization is rendered
Most visualizations use a combination of React and a visualization
library to draw charts. This makes it hard to know when the
visualization is fully rendered into the DOM from outer components,
which is needed in order to calculate the visualization size, among
others.

To solve this problem, this commit adds an optional `onRenderComplete`
callback that consumers of the visualization components may provide
to be informed when the visualization is rendered.

The new `RenderCompletionObserver` component abstracts most of the work
needed in order to get a call when the visualization is rendered for the
first time.